refactor: moved remote presentation machine
